cloud_functions/phase_07_ifc/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
ifcopenshell==0.7.*
